added pathfinding and refactored function files
