+++ client/functions/calculateAllotmentPoints.js
