+++ client/functions/divvySpaces.js
