+++ client/functions/gridGenerator.js
