+++ client/functions/index.js
@@ -1,0 +1,5 @@
+export * from './gridGenerator';
+export * from './validBoard';
+export * from './divvySpaces';
+export * from './calculateAllotmentPoints';
+export * from './hexHighlighter';
